requirements.txt
tuya-connector-python>=0.1.2
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
hypothesis>=6.0.0
pytest>=7.0.0
python-socketio>=5.9.0
python-dotenv>=1.0.0
